# src/docs/swagger.yaml
components:
    schemas:
        api.ErrorResponse:
            properties:
                error:
                    type: string
            required:
                - error
            type: object
...
    securitySchemes:
        ApiTokenAuth:
            description: Required when API token authentication is configured; otherwise no API key is required.
            in: header
            name: X-PC-Token-Key
            type: apiKey
...
info:
    contact:
        name: Process Compose Discord Channel
        url: https://discord.gg/S4xgmRSHdC
    description: This is a sample Process Compose server.
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    title: Process Compose API
    version: "1.0"
openapi: 3.0.3
paths:
...
    /process/logs/ws:
        get:
            description: Upgrades HTTP to WebSocket and streams JSON log messages. Each message is api.LogMessage.
            operationId: LogsStream
            parameters:
                - description: Comma-separated process names to stream
                  in: query
                  name: name
                  required: true
                  schema:
                    minLength: 1
                    pattern: '[^,]'
                    type: string
                - description: Offset from the end of the log
                  in: query
                  name: offset
                  required: true
                  schema:
                    type: integer
                - description: If true, continue streaming new lines
                  in: query
                  name: follow
                  schema:
                    default: false
                    type: boolean
            responses:
                "101":
                    description: Switching Protocols
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Invalid name, offset, follow, or WebSocket upgrade request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid before upgrade
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Cross-origin WebSocket upgrade request rejected
            summary: Stream process logs over WebSocket
            tags:
                - Process
...
    /process/states/ws:
        get:
            description: Upgrades HTTP to WebSocket and streams JSON events. On connect, an event with snapshot=true is sent for every (filtered) process; subsequently, every Status or Health transition (and final exit info) produces one event.
            operationId: StatesStream
            parameters:
                - description: Comma-separated process names to subscribe to. If omitted, subscribes to all processes.
                  in: query
                  name: name
                  schema:
                    type: string
            responses:
                "101":
                    description: Switching Protocols
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Invalid WebSocket upgrade request
                "401":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: API token missing or invalid before upgrade
                "403":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.ErrorResponse'
                    description: Cross-origin WebSocket upgrade request rejected
            summary: Stream process state changes over WebSocket
            tags:
                - Process
...
security:
    - ApiTokenAuth: []
    - {}
servers:
    - url: http://localhost:8080/